+++ cmd/debugger/main.go
@@ -20,8 +20,8 @@ import (
 	"fmt"
 
 	"github.com/google/exposure-notifications-server/internal/debugger"
-	"github.com/google/exposure-notifications-server/internal/logging"
 	"github.com/google/exposure-notifications-server/internal/setup"
+	"github.com/google/exposure-notifications-server/pkg/logging"
 	"github.com/google/exposure-notifications-server/pkg/server"
 	"github.com/sethvargo/go-signalcontext"
 )
