+++ cmd/generate/main.go
@@ -22,8 +22,8 @@ import (
 	"net/http"
 
 	"github.com/google/exposure-notifications-server/internal/generate"
-	"github.com/google/exposure-notifications-server/internal/logging"
 	"github.com/google/exposure-notifications-server/internal/setup"
+	"github.com/google/exposure-notifications-server/pkg/logging"
 	_ "github.com/google/exposure-notifications-server/pkg/observability"
 	"github.com/google/exposure-notifications-server/pkg/server"
 	"github.com/sethvargo/go-signalcontext"
