+++ cmd/key-rotation/main.go
@@ -20,8 +20,8 @@ import (
 	"fmt"
 
 	"github.com/google/exposure-notifications-server/internal/keyrotation"
-	"github.com/google/exposure-notifications-server/internal/logging"
 	"github.com/google/exposure-notifications-server/internal/setup"
+	"github.com/google/exposure-notifications-server/pkg/logging"
 	_ "github.com/google/exposure-notifications-server/pkg/observability"
 	"github.com/google/exposure-notifications-server/pkg/server"
 	"github.com/sethvargo/go-signalcontext"
