+++ internal/database/connection.go
@@ -21,7 +21,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/google/exposure-notifications-server/internal/logging"
+	"github.com/google/exposure-notifications-server/pkg/logging"
 
 	"github.com/jackc/pgx/v4/pgxpool"
 )
