+++ internal/revision/revision.go
@@ -28,10 +28,10 @@ import (
 	"sync"
 	"time"
 
-	"github.com/google/exposure-notifications-server/internal/logging"
 	"github.com/google/exposure-notifications-server/internal/pb"
 	"github.com/google/exposure-notifications-server/internal/publish/model"
 	"github.com/google/exposure-notifications-server/internal/revision/database"
+	"github.com/google/exposure-notifications-server/pkg/logging"
 	"google.golang.org/protobuf/proto"
 )
